Refactored import RC code to RCImporter
