+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/ui/inject/Injector.kt
@@ -2,6 +2,7 @@ package org.wycliffeassociates.otter.jvm.app.ui.inject
 
 import org.wycliffeassociates.otter.jvm.device.audio.injection.DaggerAudioComponent
 import org.wycliffeassociates.otter.jvm.device.audioplugin.injection.DaggerAudioPluginComponent
+import org.wycliffeassociates.otter.jvm.persistence.RcImporter
 import org.wycliffeassociates.otter.jvm.persistence.injection.DaggerPersistenceComponent
 import org.wycliffeassociates.otter.jvm.persistence.repositories.*
 import org.wycliffeassociates.otter.jvm.persistence.repositories.mapping.LanguageMapper
@@ -39,4 +40,9 @@ class Injector : Component(), ScopedInstance {
 
     val audioPluginRegistrar = audioPluginComponent.injectRegistrar()
 
+    val rcImporter = RcImporter(
+            database,
+            collectionRepo,
+            contentRepository
+    )
 }
